Rename module and commands to aboutcode
 * remove attributecode references everywhere
 * rename the command to aboutcode instead of about-code
