+++ NOTICE
@@ -1,4 +1,4 @@
- Copyright (c) 2013-2017 nexB Inc. http://www.nexb.com/ - All rights reserved.
+ Copyright (c) 2013-2018 nexB Inc. and others http://www.nexb.com/ - All rights reserved.
  Licensed under the Apache License, Version 2.0 (the "License");
  you may not use this file except in compliance with the License.
  You may obtain a copy of the License at
