+++ aboutcode.cfg
@@ -1,4 +1,4 @@
-# AboutCode configuration file
+# AboutCode toolkit configuration file
 
 [mappings]
 # When you generate ABOUT files from a CSV inventory, use this section to map
