+++ src/aboutcode/mapping.config
@@ -11,7 +11,7 @@
 #   * name
 #
 # Syntax:
-# <attributecode key>: <input key>
+# <aboutcode key>: <input key>
 #
 # Example:
 #   Assuming your input have columns "Resource", "Component", "Confirmed Version", "Confirmed Copyright", "Confirmed License"
